pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "starintel_doc"
authors = [
    {name = "nsaspy", email = "nsaspy@airmail.cc"},
]
description = "Starintel Documents"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Framework :: Starintel",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "py-ulid",
]
dynamic = ["version"]

